refactor heatmap into it's own package
