+++ info.limpet.ui/plugin.xml
@@ -43,7 +43,7 @@
       <view
             allowMultiple="true"
             category="info.limpet.ui"
-            class="info.limpet.ui.xy_plot.HeatmapView"
+            class="info.limpet.ui.heatmap.HeatmapView"
             icon="icons/heatmap.png"
             id="info.limpet.ui.HeatMapView"
             name="Heatmap Plot">
@@ -51,7 +51,7 @@
       <view
             allowMultiple="true"
             category="info.limpet.ui"
-            class="info.limpet.ui.xy_plot.TabularView"
+            class="info.limpet.ui.heatmap.TabularView"
             icon="icons/tabularview.png"
             id="info.limpet.ui.TabularView"
             name="Tabular view">
