+++ info.limpet.ui/src/info/limpet/ui/heatmap/TabularView.java
@@ -12,12 +12,12 @@
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  *****************************************************************************/
-package info.limpet.ui.xy_plot;
+package info.limpet.ui.heatmap;
 
 import info.limpet.IStoreItem;
 import info.limpet.impl.NumberDocument;
 import info.limpet.operations.CollectionComplianceTests;
-import info.limpet.ui.xy_plot.Helper2D.HContainer;
+import info.limpet.ui.heatmap.Helper2D.HContainer;
 
 import java.text.DecimalFormat;
 import java.util.ArrayList;
